Fix Linux and Mac build failing Github Actions (ogg).

+++ code/libogg-1.3.4/include/ogg/os_types.h
@@ -71,12 +71,12 @@
 #elif (defined(__APPLE__) && defined(__MACH__)) /* MacOS X Framework build */
 
 #  include <sys/types.h>
-   typedef int16_t ogg_int16_t;
-   typedef uint16_t ogg_uint16_t;
-   typedef int32_t ogg_int32_t;
-   typedef uint32_t ogg_uint32_t;
-   typedef int64_t ogg_int64_t;
-   typedef uint64_t ogg_uint64_t;
+   typedef signed short ogg_int16_t;
+   typedef unsigned short ogg_uint16_t;
+   typedef signed int ogg_int32_t;
+   typedef unsigned int ogg_uint32_t;
+   typedef signed long long ogg_int64_t;
+   typedef unsigned long long ogg_uint64_t;
 
 #elif defined(__HAIKU__)
